1) Updated project file for latest changes. 2) Removed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/irrlicht@4270 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/guiengine/CGUIFont.cpp
@@ -572,7 +572,7 @@ void ScalableFont::draw(const core::stringw& text, const core::rect<s32>& positi
 			offsets.push_back(offset);
 		}
 
-		offset.X += (area.width + area.overhang + GlobalKerningWidth) * m_scale;
+		offset.X += (int)((area.width + area.overhang + GlobalKerningWidth) * m_scale);
 	}
 
 	//SpriteBank->draw2DSpriteBatch(indices, offsets, clip, color);
